new game and load game done
there's some basic metadata stored, such as the date of creation and the time played. the name of the save has a number added to the end if its a duplicate automatically. also, the working directory is updated when creating or loading a save in preparation for the next commit.